requirements.txt
# Backend
fastapi
uvicorn[standard]
python-dotenv
pydantic
httpx

# LangChain + LangGraph
langchain
langchain-community
langchain-core
langgraph

# LLM - Ollama
langchain-ollama
ollama

# Embeddings
sentence-transformers

# Vector Databases
faiss-cpu
chromadb

# Repository Processing
gitpython
tree-sitter==0.21.3

# Analysis
networkx

# Frontend
streamlit

# Utilities
tqdm